мелкие правки тестов делегирования пользователям, создания старого ftl и  тесты деактивации страхового договора
