+++ tests/test_address_edit.py
@@ -98,7 +98,7 @@ def test_address_edit_lke(domain):
     add_address.backspace_all_and_input(add_address.max_height_input,
                                         base.random_value_float_str(2.0, 5.0, precision=1))
     add_address.backspace_all_and_input(add_address.max_capacity_input, base.random_value_float_str(1000, 5000))
-    add_address.dropdown_click_input_click(add_address.loading_type_select, "Боковая")
+    add_address.dropdown_click_input_click(add_address.loading_type_select, "Задняя")
     add_address.click_button(add_address.entry_pass_toggl)
     add_address.backspace_all_and_input(add_address.time_departure_input, base.random_value_float_str(10, 60))
     add_address.backspace_all_and_input(add_address.time_arrival_input, base.random_value_float_str(10, 60))
@@ -166,7 +166,7 @@ def test_address_edit_lkz(domain):
     # Ввод требовании к ТС на адресе
     add_address.input_in_field(add_address.max_height_input, base.random_value_float_str(2.0, 5.0, precision=1))
     add_address.input_in_field(add_address.max_capacity_input, base.random_value_float_str(1000, 5000))
-    add_address.dropdown_click_input_click(add_address.loading_type_select, "Верхняя")
+    add_address.dropdown_click_input_click(add_address.loading_type_select, "Боковая")
     add_address.click_button(add_address.entry_pass_toggl)
     add_address.input_in_field(add_address.time_departure_input, base.random_value_float_str(10, 60))
     add_address.input_in_field(add_address.time_arrival_input, base.random_value_float_str(10, 60))
@@ -221,7 +221,7 @@ def test_address_edit_lkz(domain):
     add_address.backspace_all_and_input(add_address.max_height_input,
                                         base.random_value_float_str(2.0, 5.0, precision=1))
     add_address.backspace_all_and_input(add_address.max_capacity_input, base.random_value_float_str(1000, 5000))
-    add_address.dropdown_click_input_click(add_address.loading_type_select, "Боковая")
+    add_address.dropdown_click_input_click(add_address.loading_type_select, "Задняя")
     add_address.click_button(add_address.entry_pass_toggl)
     add_address.backspace_all_and_input(add_address.time_departure_input, base.random_value_float_str(10, 60))
     add_address.backspace_all_and_input(add_address.time_arrival_input, base.random_value_float_str(10, 60))
